[BE-009]: Update answer handlers

+++ src/api/v1/routes/answer.py
@@ -3,6 +3,7 @@ from uuid import UUID
 
 from fastapi import APIRouter, Depends, HTTPException, Path, status
 
+from src.core.logger import logger
 from src.db.session import DBSession
 from src.domain.schemas.answer import AnswerCreate, AnswerUpdate
 from src.services.answer import AnswerService
@@ -21,7 +22,6 @@ router = APIRouter(
 )
 async def create_answer(db: DBSession, data: AnswerCreate):
     _service = AnswerService(session=db)
-
     response = await _service.create_answer(data=data)
     if not response:
         raise HTTPException(
@@ -38,7 +38,6 @@ async def create_answer(db: DBSession, data: AnswerCreate):
 )
 async def get_answer(db: DBSession, answer_id: Annotated[UUID, Path()]):
     _service = AnswerService(session=db)
-
     response = await _service.get_answer(answer_id=answer_id)
     if not response:
         raise HTTPException(
@@ -53,10 +52,17 @@ async def get_answer(db: DBSession, answer_id: Annotated[UUID, Path()]):
     status_code=status.HTTP_200_OK,
 )
 async def update_answer(
-    db: DBSession, answer_id: Annotated[UUID, Path()], data: AnswerUpdate
+    db: DBSession,
+    answer_id: Annotated[UUID, Path()],
+    data: AnswerUpdate,
+    is_answer_owner: bool = Depends(AuthService.is_answer_owner),
 ):
-    _service = AnswerService(session=db)
+    if not is_answer_owner:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = AnswerService(session=db)
     response = await _service.update_answer(answer_id=answer_id, data=data)
     if not response:
         raise HTTPException(
@@ -70,9 +76,17 @@ async def update_answer(
     "/{answer_id}/",
     status_code=status.HTTP_200_OK,
 )
-async def delete_answer(db: DBSession, answer_id: Annotated[UUID, Path()]):
-    _service = AnswerService(session=db)
+async def delete_answer(
+    db: DBSession,
+    answer_id: Annotated[UUID, Path()],
+    is_answer_owner: bool = Depends(AuthService.is_answer_owner),
+):
+    if not is_answer_owner:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Access denied"
+        )
 
+    _service = AnswerService(session=db)
     response = await _service.delete_answer(answer_id=answer_id)
     if not response:
         raise HTTPException(
